// src/models/payment.ts
export interface PayPalLink {
  href: string;
  rel: string;
  method: string;
}
// ...
export interface PayPalSaleCompletedEvent {
  id: string;
  event_version: string;
  create_time: string;
  resource_type: string;
  event_type: string;
  summary: string;
  resource: PayPalSaleResource;
  links: PayPalLink[];
}

export interface PayPalSaleResource {
  billing_agreement_id: string;
  amount: {
    total: string;
    currency: string;
    details: {
      subtotal: string;
    };
  };
  payment_mode: string;
  update_time: string;
  create_time: string;
  protection_eligibility_type: string;
  transaction_fee: {
    currency: string;
    value: string;
  };
  protection_eligibility: string;
  links: PayPalLink[];
  id: string;
  state: string;
  invoice_number: string;
}